team-5-dev/AI_BOT/requirements.txt
google-genai==0.3.0
sounddevice
numpy
scipy
faster-whisper
transformers
torch